Tool Scene 코드 통합 외 1건
CSprite 일부 변경

+++ Client/HubUIMgr.cpp
@@ -65,7 +65,7 @@ void HubUIMgr::SetBulletUI(UINT count)
 	vector<CUI*> vecBulletUI = mBulletRowUI->GetChild();
 	for (int i = 0; i < vecBulletUI.size(); ++i)
 	{
-		if (i >= count)
+		if (i >= (int)count)
 		{
 			vecBulletUI[i]->SetVisible(false);
 		}
